Refactor: move metadata components under widgets and datasets folders. Minor eslint fixes also included
